requirements.txt
asttokens==2.0.5
backcall==0.2.0
boto3==1.21.33
botocore==1.24.33
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
fonttools==4.31.2
greenlet==1.1.2
ipykernel==6.12.1
ipython==8.2.0
ipython-genutils==0.2.0
ipython-sql==0.4.0
jedi==0.18.1
jmespath==1.0.0
jupyter-client==7.2.1
jupyter-core==4.9.2
kiwisolver==1.4.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.3
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prettytable==0.7.2
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
pyzmq==22.3.0
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.34
sqlparse==0.4.2
stack-data==0.2.0
tabulate==0.8.9
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
